+++ src/Exception/UnauthorizedException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Icawebdesign\Hibp\Exception;
+
+use RuntimeException;
+
+class UnauthorizedException extends RuntimeException
+{
+
+}