+++ src/Subscription/Subscription.php
@@ -7,6 +7,7 @@ use GuzzleHttp\ClientInterface;
 use Icawebdesign\Hibp\Traits\HibpConfig;
 use GuzzleHttp\Exception\ClientException;
 use GuzzleHttp\Exception\RequestException;
+use Icawebdesign\Hibp\Exception\UnauthorizedException;
 use Icawebdesign\Hibp\Exception\PasteNotFoundException;
 
 use function sprintf;
@@ -25,7 +26,7 @@ class Subscription implements SubscriptionInterface
 
     public function __construct(HibpHttp $hibpHttp)
     {
-        $this->apiRoot = "{$this->hibp['api_root']}/v/{$this->hibp['api_version']}/";
+        $this->apiRoot = "{$this->hibp['api_root']}/v{$this->hibp['api_version']}";
         $this->client = $hibpHttp->client();
     }
 
@@ -41,7 +42,7 @@ class Subscription implements SubscriptionInterface
             $this->statusCode = $exception->getCode();
 
             throw match ($exception->getCode()) {
-                404 => new PasteNotFoundException($exception->getMessage()),
+                401 => new UnauthorizedException($exception->getMessage(), $exception->getCode()),
                 400 => new RequestException($exception->getMessage(), $exception->getRequest()),
                 default => $exception,
             };